backend/requirements.txt
camel-ai[all]
python-dotenv
requests
Pillow
urllib3
